+++ go.mod
@@ -1,0 +1,8 @@
+module github.com/lyvecloud/vault-exfiltrate
+
+go 1.15
+
+require (
+	github.com/hashicorp/vault v1.8.1
+	golang.org/x/exp v0.0.0-20210729172720-737cce5152fc
+)
